% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_nodes}
\alias{get_nodes}
\title{Extract nodes from R code}
\usage{
get_nodes(path_to_file, ignore_source = NULL)
}
\arguments{
\item{path_to_file}{character: A file path string}

\item{ignore_source}{character: If not NULL, any file names matching those
specified will be ignored in any "source(filename)" within the scripts.}
}
\value{
Returns a dataframe with columns id, assign, member, function, code
}
\description{
Extract nodes from R code
}
