% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{make_dot}
\alias{make_dot}
\title{Convert nodes/edges into a dotfile format for dataflow graph}
\usage{
make_dot(flow)
}
\arguments{
\item{flow}{list of nodes/edges returned by \code{\link{get_flow}}}
}
\value{
Returns a string in a dotfile-compatible format
}
\description{
Convert nodes/edges into a dotfile format for dataflow graph
}
