% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{parameterize_flow}
\alias{parameterize_flow}
\title{Prepare flow (nodes, edges) for graphing}
\usage{
parameterize_flow(flow, label_option = "auto", hover_code = "node")
}
\arguments{
\item{flow}{list of nodes/edges returned by \code{\link{get_flow}}}

\item{label_option}{character: Either "both", "assign", "function" or
"auto" (which uses "assign" for input nodes and "function" for others).}

\item{hover_code}{character: Either "node", "network", or NULL. If "node", code
for the current node will be displayed on hover, all dependent code for
"network", and none if NULL.}
}
\value{
Returns a modified list of nodes/edges
}
\description{
Prepare flow (nodes, edges) for graphing
}
