% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{prep_flow}
\alias{prep_flow}
\title{Prepare flow (nodes, edges) for graphing}
\usage{
prep_flow(
  flow,
  focus_node = NULL,
  prune_labels = NULL,
  prune_all_functions = TRUE,
  prune_all_mutates = FALSE,
  label_option = "auto",
  hover_code = "node"
)
}
\arguments{
\item{flow}{list of nodes/edges returned by \code{\link{get_flow}}}

\item{focus_node}{numeric: ID of a node to focus upon. Only the dependencies
of the specified node will be shown.}

\item{prune_labels}{character: If not NULL, any nodes with labels matching
those specified will be pruned (i.e., excluded from the plot) although their
dependencies will be cascaded.}

\item{prune_all_functions}{logical: If TRUE, assigned function nodes will be
pruned from the plot.}

\item{prune_all_mutates}{logical: If TRUE, intermediate mutated nodes (those
with only a self-dependency) will pruned from the plot.}

\item{label_option}{character: Either "both", "assign", "function" or
"auto" (which uses "assign" for input nodes and "function" for others).}

\item{hover_code}{character: Either "node", "network", or NULL. If "node", code
for the current node will be displayed on hover, all dependent code for
"network", and none if NULL.}
}
\value{
Returns a modified list of nodes/edges
}
\description{
Prepare flow (nodes, edges) for graphing
}
