% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{prune_flow}
\alias{prune_flow}
\title{Identify nodes to prune}
\usage{
prune_flow(
  flow,
  focus_node = NULL,
  prune_labels = NULL,
  prune_all_functions = TRUE,
  prune_all_mutates = FALSE
)
}
\arguments{
\item{flow}{list of nodes/edges returned by \code{\link{get_flow}}}

\item{focus_node}{numeric: ID of a node to focus upon. Only the dependencies
of the specified node will be shown.}

\item{prune_labels}{character: If not NULL, any nodes with labels matching
those specified will be pruned (i.e., excluded from the plot) although their
dependencies will be cascaded.}

\item{prune_all_functions}{logical: If TRUE, assigned function nodes will be
pruned from the plot.}

\item{prune_all_mutates}{logical: If TRUE, intermediate mutated nodes (those
with only a self-dependency) will pruned from the plot.}
}
\value{
Returns list of nodes/edges and (optionally) pruned_ids
}
\description{
Identify nodes to prune
}
