% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{get_edges}
\alias{get_edges}
\title{Extract edges from nodes}
\usage{
get_edges(nodes)
}
\arguments{
\item{nodes}{dataframe returned by \code{\link{get_nodes}}}
}
\value{
Returns a dataframe with two columns (to and from) that each hold node IDs
}
\description{
Extract edges from nodes
}
