% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{plot_flow}
\alias{plot_flow}
\title{Plot dataflow graph from R code}
\usage{
plot_flow(
  path_to_file,
  focus_node = NULL,
  prune_labels = NULL,
  prune_all_functions = TRUE,
  prune_all_mutates = FALSE,
  label_option = "auto",
  hover_code = TRUE,
  ignore_source = NULL
)
}
\arguments{
\item{path_to_file}{character: A file path string}

\item{focus_node}{numeric: ID of a node to focus upon. Only the dependencies
of the specified node will be shown.}

\item{prune_labels}{character: If not NULL, any nodes with labels matching
those specified will be pruned (i.e., excluded from the plot) although their
dependencies will be cascaded.}

\item{prune_all_functions}{logical: If TRUE, assigned function nodes will be
pruned from the plot.}

\item{prune_all_mutates}{logical: If TRUE, intermediate mutated nodes (those
with only a self-dependency) will pruned from the plot.}

\item{label_option}{character: Either "both", "assign", "function" or
"auto" (which uses "assign" for input nodes and "function" for others).}

\item{hover_code}{logical: If TRUE, code (for a node) will be displayed on
hover. Code for pruned nodes will be displayed as part of the hovered text
for their immediate downstream nodes.}

\item{ignore_source}{character: If not NULL, any file names matching those
specified will be ignored in any "source(filename)" within the scripts.}
}
\value{
Returns a data flow diagram rendered by \code{\link[DiagrammeR]{grViz}}
}
\description{
Plot dataflow graph from R code
}
