% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{plot_flow}
\alias{plot_flow}
\title{Plot dataflow graph from R code}
\usage{
plot_flow(path_to_file, collapse_nodes = TRUE, exclude_text = FALSE)
}
\arguments{
\item{path_to_file}{character: A file path string}

\item{collapse_nodes}{logical: If TRUE, any dependencies on "mutate" nodes
will be set to their parent "non-mutate" nodes.}

\item{exclude_text}{logical: If TRUE, code for a node will not be available
on hover}
}
\value{
Returns a data flow diagram rendered by \code{\link[DiagrammeR]{grViz}}
}
\description{
Plot dataflow graph from R code
}
