% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{plot_flow}
\alias{plot_flow}
\title{Plot dataflow graph from R code}
\usage{
plot_flow(
  path_to_file,
  ignore_source = NULL,
  prune_labels = NULL,
  prune_all_functions = TRUE,
  label_option = "both",
  exclude_text = FALSE
)
}
\arguments{
\item{path_to_file}{character: A file path string}

\item{ignore_source}{character: If not NULL, any file names matching those
specified will be ignored in any "source(filename)" within the scripts.}

\item{prune_labels}{character: If not NULL, any nodes with labels matching
those specified will be dropped (although their dependencies will be cascaded).}

\item{prune_all_functions}{logical: If TRUE, assinged functions will be pruned
from the plot}

\item{label_option}{character: Either "both" (default), "assign", "function" or
"auto" (which uses "assign" for input nodes and "function" for others).}

\item{exclude_text}{logical: If TRUE, code for a node will not be available
on hover}
}
\value{
Returns a data flow diagram rendered by \code{\link[DiagrammeR]{grViz}}
}
\description{
Plot dataflow graph from R code
}
